refactor(nef): extract package-graph module
Move PackageGraph and its dependency-argument resolution
(try_resolve_dependency_argument, read_and_analyze) to scan/graph.rs.
The module borrows FileInfo, analyze_file_at, and identity_origins from
its parent via 'use super'.
